controller_note now finds controllers in engines
With rails engines controllers can be located in
plugins.

If ActionController::Routing responds to the message
:controller_paths then this patch loops through all
the paths to controllers and returns the first one
where the file referenced by the full path exists.

Otherwise the controller_filename method uses the
previous method for generating the path for the
controller.

Signed-off-by: José Valim <jose.valim@gmail.com>