Apparently, Filter.prefix is pretty strict on the appearance order of %s, %d and %d for the file name, line number and column number in that the %s has to be appear first in the string before any of the %d. Otherwise, you get
    Footnotes #<Footnotes::Notes::ControllerNote:0x9a5a650> Exception: invalid value for Integer(): "/opt/ror/tribes/app/controllers/dailies_controller.rb"

errors all over the place. I ran into this when I tried writing my own gedit protocol wrapper where I assumed the format to be:

    foo://l=%d&%s

and ran into problems.

Granted that I'm probably an edge case and not a lot of people write their own filter prefixes so this is probably not worth writing code for. Just wanted it to appear in the docs somewhere to spare someone else the trouble I've gone through.